requirements.txt
asgiref==3.7.2
Django==3.2.16
django-bootstrap5==23.4
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
pytils==0.4.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
